+++ packages/types/src/querier.ts
@@ -9,9 +9,15 @@ import {
   DutchAuction,
 } from '@penumbra-zone/protobuf/penumbra/core/component/auction/v1/auction_pb';
 import { CompactBlock } from '@penumbra-zone/protobuf/penumbra/core/component/compact_block/v1/compact_block_pb';
+import {
+  CurrentGasPricesRequest,
+  CurrentGasPricesResponse,
+} from '@penumbra-zone/protobuf/penumbra/core/component/fee/v1/fee_pb';
 import { LqtCheckNullifierResponse } from '@penumbra-zone/protobuf/penumbra/core/component/funding/v1/funding_pb';
 import {
   Nullifier,
+  SctFrontierRequest,
+  SctFrontierResponse,
   TimestampByHeightRequest,
   TimestampByHeightResponse,
 } from '@penumbra-zone/protobuf/penumbra/core/component/sct/v1/sct_pb';
@@ -85,8 +91,13 @@ export interface AuctionQuerierInterface {
 
 export interface SctQuerierInterface {
   timestampByHeight(req: TimestampByHeightRequest): Promise<TimestampByHeightResponse>;
+  sctFrontier(req: SctFrontierRequest): Promise<SctFrontierResponse>;
 }
 
 export interface FundingQuerierInterface {
   lqtCheckNullifier(epochIndex: bigint, nullifier: Nullifier): Promise<LqtCheckNullifierResponse>;
+}
+
+export interface FeeQuerierInterface {
+  currentGasPrices(req: CurrentGasPricesRequest): Promise<CurrentGasPricesResponse>;
 }