+++ packages/getters/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @penumbra-zone/getters
 
+## 27.0.0
+
+### Patch Changes
+
+- Updated dependencies [dc1eb8b]
+  - @penumbra-zone/protobuf@10.1.0
+  - @penumbra-zone/bech32m@17.0.0
+
 ## 26.0.0
 
 ### Patch Changes