+++ packages/ui-deprecated/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @penumbra-zone/ui
 
+## 20.0.0
+
+### Patch Changes
+
+- Updated dependencies [dc1eb8b]
+- Updated dependencies [f9cd9dd]
+  - @penumbra-zone/protobuf@10.1.0
+  - @penumbra-zone/types@34.0.0
+  - @penumbra-zone/bech32m@17.0.0
+  - @penumbra-zone/perspective@57.0.0
+
+## 19.0.1
+
+### Patch Changes
+
+- Updated dependencies [085e855]
+  - @penumbra-zone/types@33.1.0
+  - @penumbra-zone/perspective@56.0.0
+
 ## 19.0.0
 
 ### Major Changes